Fix build with the latest version of Node 12
A recent release of Node 12 required that exports have explicit file
extensions which was patched in babel/babel#11283. Ran the following
commands to bump the dependencies and had to add `style-loader`.

    yarn upgrade @storybook/react
    yarn add style-loader

After running these the package successfully built on Node 12.21.0.

of @storybook/react in yarn.lock so that we pick up this change.